Working jupytext conversion.
